continue moving functions to styler, adjusting styler functions some need dlg in parameters after split
